+++ backend/models/coursesDAO.js
@@ -1,10 +1,10 @@
 const db = require('../src/database');
 
 const sql = {
-  courseList: `SELECT c.course_id, u.nickname, title, attach_image_path, DATE_FORMAT(c.createdAt, '%Y-%m-%d %H:%i') as createdAt
+  courseList: `SELECT c.course_id, u.nickname, title, attach_image_path, view_cnt, DATE_FORMAT(c.createdAt, '%Y-%m-%d %H:%i') as createdAt
                FROM users u INNER JOIN courses c ON u.user_id = c.user_id
                ORDER BY c.course_id DESC;`, // 강의번호로 내림차순 (GROUP BY c.course_id 해야하나?)
-  course: `SELECT c.user_id, c.course_id, u.nickname, title, content, category, attach_file_path, attach_file_name, attach_image_path, DATE_FORMAT(c.createdAt, '%Y-%m-%d') as createdAt
+  course: `SELECT c.user_id, c.course_id, u.nickname, title, content, category, attach_file_path, attach_file_name, attach_image_path, view_cnt, DATE_FORMAT(c.createdAt, '%Y-%m-%d') as createdAt
             FROM users u INNER JOIN courses c ON u.user_id = c.user_id
             WHERE c.course_id = ?;`, // 특정 강의 상세조회
   insert: `INSERT INTO courses(title, content, category, user_id, attach_file_path, attach_file_name, attach_image_path) 
@@ -15,18 +15,13 @@ const sql = {
   search: `SELECT c.course_id, u.nickname, title, attach_image_path, DATE_FORMAT(c.createdAt, '%Y-%m-%d %H:%i') as createdAt
            FROM users u INNER JOIN courses c ON u.user_id = c.user_id
            WHERE title LIKE CONCAT('%', ?, '%') OR content LIKE CONCAT('%', ?, '%');`,
+  increase_view: `UPDATE courses SET view_cnt = view_cnt + 1 WHERE course_id = ?`,
 };
 
 const coursesDAO = {
   courseList: async (callback) => {
-    // const no = Number(item.no) - 1 || 0;
-    // const size = Number(item.size) || 10;
-
     try {
       const [resdata] = await db.query(sql.courseList);
-      // const [resp] = await db.query(sql.courseList, [no * size, size]);
-      // const [cntResp] = await db.query(sql.totalCount);
-      // const totalPage = Math.ceil(cntResp[0].cnt / size);
       callback({
         status: 200,
         message: '강의 리스트 조회 성공',
@@ -40,6 +35,7 @@ const coursesDAO = {
 
   course: async (id, callback) => {
     try {
+      await db.query(sql.increase_view, [id]);
       const resp = await db.query(sql.course, [id]);
       if (resp[0].length === 0) {
         // course_id에 해당하는 강의가 없다면
